L19 13:14

+++ Lektion-19/eShop-solution/api/DTOs/Orders/PostDeliveryMethodDto.cs
@@ -11,5 +11,5 @@ public class PostDeliveryMethodDto
     [Required]
     public required string Description { get; set; }
     [Required]
-    public double Price { get; set; } = 0.00;
+    public decimal Price { get; set; } = 0.00M;
 }
